debt(AB#00): Updates git hook msg lint task to config with project init

+++ app/src/test/java/com/downstairs/tosplit/split/list/SplitsViewModelTest.kt
@@ -102,7 +102,7 @@ class SplitsViewModelTest {
     }
 
     private fun getSplit(
-        id: Int = 1,
+        id: Long = 1,
         name: String = "Car rent",
         user: UserRemote = UserRemote("Some Payer"),
         value: Double = 230.00
